chore(build-wysiwyg): PR 8b — delete dead grid components
PR 8a switched the page render to BuildWysiwyg. This PR removes the now-
unreachable grid components, their tests, and the mobile-grid variants.

Deleted:
- BuildGrid + index, GridHeader, GridBody, GroupedBody, ScrollableTable,
  SideRail, FieldEditor, Toolbar, CellInput, CapacityCell, ResizeHandle
- columnSizing, fieldEditorValidation (grid-only utilities)
- mobile/MobileBuildGrid, SlotList, SlotCard, FieldChipStrip,
  MobileGroupByControl, MobileBottomBar, MobilePreviewSheet
- All corresponding `*.test.tsx` / `*.test.ts` files

Kept (still consumed by build-wysiwyg/):
- useGridState (hook + reducer)
- useReorderable (hook)
- fieldTypes (icon + label meta)
- SaveStatus (the badge component)

The `SignupMeta` type previously lived in BuildGrid.tsx; it's now an
export from BuildWysiwyg.tsx since that's the only remaining consumer.

The `build-grid/` folder name is misleading now that it only holds
shared hooks. Leaving it as-is here to keep the deletion PR pure; a
follow-up can rename to `build-shared/` or similar without functional
change.

453 unit tests pass (down from 528 — the dropped 75 were grid-specific
tests we removed alongside their components); lint + typecheck clean.
Chrome MCP smoke at /app/signups/[id]/build confirms the WYSIWYG sheet
still renders with no errors.
