--- src/components/build-grid/index.ts
@@ -1,1 +1,0 @@
-export { BuildGrid } from './BuildGrid';
